# Further md5crypt-opencl optimizations