@solardiz 2019-02-02
Got further optimizations, significantly outperforming Hashcat on Titan X Maxwell. Not sure if we want to merge this PR just yet. Maybe I'll add commits and edit the description.