@solardiz 2019-02-02
Pushed another commit, will update the description with new benchmarks separately. There are no ill effects except that optimal GWS became even larger. I'm thinking of also adding detection of different-length passwords via local memory and disabling the per-length specialization if so.