@solardiz 2019-02-02
Pushed yet another commit, disabling specialization to password length if it varies within the group. This has a performance impact of maybe 2% when the length does not vary. It's also much less optimal than not even having the specialization code compiled in when the length does very. But it feels like a reasonable trade-off for now.

We have worse issues with the auto-tuning anyway, where it tunes based on test vectors that might not reflect further candidate passwords.