@solardiz 2019-02-03
I realized that my commit "disabling specialization to password length if it varies within the group" effectively excludes large portions of code from our self-tests. :-( Perhaps we need to add logic to avoid "disabling specialization" (at least sometimes) when running self-tests - e.g., check for a known self-test vector within the group?