@solardiz 2019-02-04
I tried fixing the presumed self-tests code coverage issue with the changes below, but somehow am unable to trigger the problem - deliberately breaking code paths that I thought were not reached in the self-test because of the differing lengths triggering `unify = 0xff;` in the currently committed code actually results in self-test failure even without the fix below, I don't know why.

```diff
 /*
- * Disable specialization to password length if it varies within the group.
+ * Disable specialization to password length if it varies within the group,
+ * except if we're the first group and there's a zero-length password in it
+ * (suggesting we're in self-test and want to maximize our code coverage).
  * We could use tree-like log2(N) reduction here, but we don't bother.
  */
 	__local uint lengths[0x100];
@@ -629,10 +631,22 @@
 	barrier(CLK_LOCAL_MEM_FENCE);
 
 	uint unify = 0;
-	for (i = 0; i < lws; i++) {
-		if (pass_len != lengths[i]) {
-			unify = 0xff;
-			break;
+	if (lid != idx) {
+		for (i = 0; i < lws; i++) {
+			if (pass_len != lengths[i]) {
+				unify = 0xff;
+				break;
+			}
+		}
+	} else {
+		for (i = 0; i < lws; i++) {
+			uint leni = lengths[i];
+			if (pass_len != leni)
+				unify = 0xff;
+			if (!leni) {
+				unify = 0;
+				break;
+			}
 		}
 	}
```